+++ js/data.js
@@ -1,0 +1,41 @@
+export const NAMES = [
+  'Артём',
+  'Мария',
+  'Дмитрий',
+  'Анна',
+  'Сергей',
+  'Елена',
+  'Алексей',
+  'Ольга',
+  'Иван',
+  'Наталья'
+];
+
+export const MESSAGES = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!'
+];
+
+export const DESCRIPTIONS = [
+  'Красивый закат на море',
+  'Городские огни ночью',
+  'Природа в её лучшем проявлении',
+  'Момент из путешествия',
+  'Уютный домашний вечер',
+  'Уличное искусство',
+  'Вкусный ужин',
+  'Спортивные достижения',
+  'Утренний кофе',
+  'Зимняя сказка'
+];
+
+export const AVATAR_COUNT = 6;
+export const MIN_LIKES = 15;
+export const MAX_LIKES = 200;
+export const MIN_COMMENTS = 0;
+export const MAX_COMMENTS = 30;
+export const PHOTOS_COUNT = 25;
